+++ packages/code-du-travail-frontend/src/lib/matomo/index.ts
@@ -1,0 +1,2 @@
+export * from "./question";
+export * from "./types";