+++ packages/code-du-travail-frontend/src/outils/ConventionCollective/enterprise/EnterpriseButton.tsx
@@ -1,79 +1,65 @@
-import { SOURCES } from "@socialgouv/cdtn-sources";
 import { Tag, Text, theme } from "@socialgouv/cdtn-ui";
-import Link from "next/link";
 import React from "react";
 import styled from "styled-components";
 
 import Html from "../../../common/Html";
 import { Enterprise } from "../../../conventions/Search/api/enterprises.service";
-import { ScreenType } from "../common/NavContext";
 import { ResultItem } from "../common/ResultList";
-import { useTrackingContext } from "../common/TrackingContext";
+import { TrackingProps, UserAction } from "../types";
 
 type CompagnyItemProps = {
   enterprise: Enterprise;
   isFirst: boolean;
   showAddress: boolean;
   onClick: (enterprise: Enterprise) => void;
-};
+} & TrackingProps;
 
 export function EnterpriseButton({
   enterprise,
   isFirst,
   showAddress,
   onClick,
+  onUserAction,
 }: CompagnyItemProps): JSX.Element {
   const {
     label,
-    etablissements,
+    matching,
     highlightLabel,
     simpleLabel,
     activitePrincipale,
     address,
     siren,
-    matchingEtablissement,
+    firstMatchingEtablissement,
   } = enterprise;
 
-  const { trackEvent, title, uuid } = useTrackingContext();
-
   const clickHandler = () => {
-    trackEvent(
-      "enterprise_select",
-      title,
-      JSON.stringify({ label, siren }),
-      uuid
-    );
+    onUserAction(UserAction.SelectEnterprise, { label, siren });
     onClick(enterprise);
   };
   const showTitleWithHighlight = label === simpleLabel;
   return (
-    <Link
-      href={`/${SOURCES.TOOLS}/convention-collective#${ScreenType.agreementSelection}`}
-      passHref
-    >
-      <ItemButton isFirst={isFirst} onClick={clickHandler}>
-        {showTitleWithHighlight ? (
-          <Title fontSize="hsmall" fontWeight="600">
-            {highlightLabel}
+    <ItemButton isFirst={isFirst} onClick={clickHandler}>
+      {showTitleWithHighlight ? (
+        <Title fontSize="hsmall" fontWeight="600">
+          {highlightLabel}
+        </Title>
+      ) : (
+        <>
+          <Title fontWeight="600" fontSize="hsmall">
+            {simpleLabel}
           </Title>
-        ) : (
-          <>
-            <Title fontWeight="600" fontSize="hsmall">
-              {simpleLabel}
-            </Title>
-            <Subtitle fontSize="small">{highlightLabel}</Subtitle>
-          </>
-        )}
-        {activitePrincipale && (
-          <Activity as="div">Activité : {activitePrincipale}</Activity>
-        )}
-        {!showAddress && etablissements > 1 ? (
-          <Tag> {etablissements} établissements </Tag>
-        ) : (
-          <Text>{address || matchingEtablissement.address}</Text>
-        )}
-      </ItemButton>
-    </Link>
+          <Subtitle fontSize="small">{highlightLabel}</Subtitle>
+        </>
+      )}
+      {activitePrincipale && (
+        <Activity as="div">Activité : {activitePrincipale}</Activity>
+      )}
+      {!showAddress && matching > 1 ? (
+        <Tag> {matching} établissements </Tag>
+      ) : (
+        <Text>{address || firstMatchingEtablissement?.address}</Text>
+      )}
+    </ItemButton>
   );
 }
 
@@ -92,6 +78,7 @@ const ItemButton = styled(ResultItem)`
 
 const Title = styled(Html)`
   padding-bottom: ${spacings.xsmall};
+
   & b {
     font-weight: 700;
   }