+++ packages/code-du-travail-frontend/src/outils/HeuresRechercheEmploi/__tests__/index.test.tsx
@@ -1,0 +1,13 @@
+import { render } from "@testing-library/react";
+import React from "react";
+
+import { HeuresRechercheEmploi } from "..";
+
+describe("<HeuresRechercheEmploi />", () => {
+  it("should render", () => {
+    const { container } = render(
+      <HeuresRechercheEmploi title="Heures pour recherche d'emploi" icon="" />
+    );
+    expect(container).toMatchSnapshot();
+  });
+});